requirements.txt
aiofiles==23.2.1 ; python_version >= "3.10" and python_version < "3.13"
altair==5.1.1 ; python_version >= "3.10" and python_version < "3.13"
annotated-types==0.5.0 ; python_version >= "3.10" and python_version < "3.13"
anyio==3.7.1 ; python_version >= "3.10" and python_version < "3.13"
attrs==23.1.0 ; python_version >= "3.10" and python_version < "3.13"
certifi==2023.7.22 ; python_version >= "3.10" and python_version < "3.13"
charset-normalizer==3.2.0 ; python_version >= "3.10" and python_version < "3.13"
click==8.1.7 ; python_version >= "3.10" and python_version < "3.13"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "3.13" and platform_system == "Windows"
contourpy==1.1.1 ; python_version >= "3.10" and python_version < "3.13"
cycler==0.11.0 ; python_version >= "3.10" and python_version < "3.13"
docutils==0.20.1 ; python_version >= "3.10" and python_version < "3.13"
exceptiongroup==1.1.3 ; python_version >= "3.10" and python_version < "3.11"
fastapi==0.103.1 ; python_version >= "3.10" and python_version < "3.13"
ffmpy==0.3.1 ; python_version >= "3.10" and python_version < "3.13"
filelock==3.12.4 ; python_version >= "3.10" and python_version < "3.13"
fonttools==4.42.1 ; python_version >= "3.10" and python_version < "3.13"
fsspec==2023.9.1 ; python_version >= "3.10" and python_version < "3.13"
gradio-client==0.5.0 ; python_version >= "3.10" and python_version < "3.13"
gradio==3.44.3 ; python_version >= "3.10" and python_version < "3.13"
h11==0.14.0 ; python_version >= "3.10" and python_version < "3.13"
httpcore==0.18.0 ; python_version >= "3.10" and python_version < "3.13"
httpx==0.25.0 ; python_version >= "3.10" and python_version < "3.13"
huggingface-hub==0.17.1 ; python_version >= "3.10" and python_version < "3.13"
idna==3.4 ; python_version >= "3.10" and python_version < "3.13"
importlib-resources==6.0.1 ; python_version >= "3.10" and python_version < "3.13"
jinja2==3.1.2 ; python_version >= "3.10" and python_version < "3.13"
jsonschema-specifications==2023.7.1 ; python_version >= "3.10" and python_version < "3.13"
jsonschema==4.19.0 ; python_version >= "3.10" and python_version < "3.13"
kiwisolver==1.4.5 ; python_version >= "3.10" and python_version < "3.13"
markupsafe==2.1.3 ; python_version >= "3.10" and python_version < "3.13"
matplotlib==3.8.0 ; python_version >= "3.10" and python_version < "3.13"
monai==1.2.0 ; python_version >= "3.10" and python_version < "3.13"
mpmath==1.3.0 ; python_version >= "3.10" and python_version < "3.13"
networkx==3.1 ; python_version >= "3.10" and python_version < "3.13"
numpy==1.26.0 ; python_version >= "3.10" and python_version < "3.13"
opencv-python==4.8.0.76 ; python_version >= "3.10" and python_version < "3.13"
orjson==3.9.7 ; python_version >= "3.10" and python_version < "3.13"
packaging==23.1 ; python_version >= "3.10" and python_version < "3.13"
pandas==2.1.0 ; python_version >= "3.10" and python_version < "3.13"
pendulum==2.1.2 ; python_version >= "3.10" and python_version < "3.13"
pillow==10.0.1 ; python_version >= "3.10" and python_version < "3.13"
pydantic-core==2.6.3 ; python_version >= "3.10" and python_version < "3.13"
pydantic==2.3.0 ; python_version >= "3.10" and python_version < "3.13"
pydub==0.25.1 ; python_version >= "3.10" and python_version < "3.13"
pyparsing==3.1.1 ; python_version >= "3.10" and python_version < "3.13"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "3.13"
python-multipart==0.0.6 ; python_version >= "3.10" and python_version < "3.13"
pytz==2023.3.post1 ; python_version >= "3.10" and python_version < "3.13"
pytzdata==2020.1 ; python_version >= "3.10" and python_version < "3.13"
pyyaml==6.0.1 ; python_version >= "3.10" and python_version < "3.13"
referencing==0.30.2 ; python_version >= "3.10" and python_version < "3.13"
requests==2.31.0 ; python_version >= "3.10" and python_version < "3.13"
rpds-py==0.10.3 ; python_version >= "3.10" and python_version < "3.13"
safetensors==0.3.3 ; python_version >= "3.10" and python_version < "3.13"
semantic-version==2.10.0 ; python_version >= "3.10" and python_version < "3.13"
setuptools-scm==7.1.0 ; python_version >= "3.10" and python_version < "3.13"
setuptools==68.2.2 ; python_version >= "3.10" and python_version < "3.13"
six==1.16.0 ; python_version >= "3.10" and python_version < "3.13"
sniffio==1.3.0 ; python_version >= "3.10" and python_version < "3.13"
starlette==0.27.0 ; python_version >= "3.10" and python_version < "3.13"
statistics==1.0.3.5 ; python_version >= "3.10" and python_version < "3.13"
sympy==1.12 ; python_version >= "3.10" and python_version < "3.13"
tomli==2.0.1 ; python_version >= "3.10" and python_version < "3.11"
toolz==0.12.0 ; python_version >= "3.10" and python_version < "3.13"
torch==2.0.1 ; python_version >= "3.10" and python_version < "3.13"
torchvision==0.15.2 ; python_version >= "3.10" and python_version < "3.13"
tqdm==4.66.1 ; python_version >= "3.10" and python_version < "3.13"
typing-extensions==4.8.0 ; python_version >= "3.10" and python_version < "3.13"
tzdata==2023.3 ; python_version >= "3.10" and python_version < "3.13"
urllib3==2.0.4 ; python_version >= "3.10" and python_version < "3.13"
uvicorn==0.23.2 ; python_version >= "3.10" and python_version < "3.13"
websockets==11.0.3 ; python_version >= "3.10" and python_version < "3.13"
